requirements.txt
pandas
flask
Jinja2
beautifulsoup4
psx-data-reader
prophet